vkd3d: Use vkd3d_atomic_increment_u32() in d3d12_fence_incref().

+++ libs/vkd3d/vkd3d_private.h
@@ -621,7 +621,7 @@ struct vkd3d_signaled_semaphore
 struct d3d12_fence
 {
     ID3D12Fence1 ID3D12Fence1_iface;
-    LONG internal_refcount;
+    unsigned int internal_refcount;
     unsigned int refcount;
 
     D3D12_FENCE_FLAGS flags;
